backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.0
pydantic==1.10.15
python-multipart==0.0.9
